Move ShaderObject to be under renderer interface.
