+++ .gitignore
@@ -7,6 +7,7 @@
 *.sdf
 *.ilk
 *.obj
+*.slang-module
 .clang-format
 bin/
 intermediate/
