+++ tools/gfx/vulkan/render-vk.cpp
@@ -3,6 +3,7 @@
 
 //WORKING:#include "options.h"
 #include "../render.h"
+#include "../render-graphics-common.h"
 
 #include "../../source/core/slang-smart-pointer.h"
 
@@ -29,7 +30,7 @@
 namespace gfx {
 using namespace Slang;
 
-class VKRenderer : public Renderer
+class VKRenderer : public GraphicsAPIRenderer
 {
 public:
     enum
@@ -62,8 +63,8 @@ public:
     Result createDescriptorSet(DescriptorSetLayout* layout, DescriptorSet** outDescriptorSet) override;
 
     Result createProgram(const ShaderProgram::Desc& desc, ShaderProgram** outProgram) override;
-    Result createGraphicsPipelineState(const GraphicsPipelineStateDesc& desc, PipelineState** outState) override;
-    Result createComputePipelineState(const ComputePipelineStateDesc& desc, PipelineState** outState) override;
+    Result createGraphicsPipelineState(GraphicsPipelineStateDesc& desc, PipelineState** outState) override;
+    Result createComputePipelineState(ComputePipelineStateDesc& desc, PipelineState** outState) override;
 
     virtual SlangResult captureScreenSurface(Surface& surface) override;
 
@@ -2721,8 +2722,10 @@ Result VKRenderer::createProgram(const ShaderProgram::Desc& desc, ShaderProgram*
     return SLANG_OK;
 }
 
-Result VKRenderer::createGraphicsPipelineState(const GraphicsPipelineStateDesc& desc, PipelineState** outState)
+Result VKRenderer::createGraphicsPipelineState(GraphicsPipelineStateDesc& desc, PipelineState** outState)
 {
+    preparePipelineDesc(desc);
+
     VkPipelineCache pipelineCache = VK_NULL_HANDLE;
 
     auto programImpl = (ShaderProgramImpl*) desc.program;
@@ -2849,8 +2852,10 @@ Result VKRenderer::createGraphicsPipelineState(const GraphicsPipelineStateDesc& 
     return SLANG_OK;
 }
 
-Result VKRenderer::createComputePipelineState(const ComputePipelineStateDesc& desc, PipelineState** outState)
+Result VKRenderer::createComputePipelineState(ComputePipelineStateDesc& desc, PipelineState** outState)
 {
+    preparePipelineDesc(desc);
+
     VkPipelineCache pipelineCache = VK_NULL_HANDLE;
 
     auto programImpl = (ShaderProgramImpl*) desc.program;
